Backend/requirements.txt
aiohttp==3.7.4.post0
aioredis==1.3.1
asgiref==3.4.1
async-timeout==3.0.1
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
beautifulsoup4==4.10.0
cffi==1.14.6
channels==3.0.4
channels-redis==3.3.0
chardet==4.0.0
constantly==15.1.0
cryptography==3.4.8
daphne==3.0.2
Django==3.2.7
django-debug-toolbar==3.2.2
djangorestframework==3.12.4
hiredis==2.0.0
hyperlink==21.0.0
idna==3.2
incremental==21.3.0
msgpack==1.0.2
multidict==5.1.0
Pillow==8.3.2
psycopg2-binary==2.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyOpenSSL==20.0.1
python-dotenv==0.19.0
pytz==2021.1
service-identity==21.1.0
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.2
Twisted==21.7.0
twisted-iocpsupport==1.0.2
txaio==21.2.1
typing-extensions==3.10.0.2
yarl==1.6.3
zope.interface==5.4.0

requests~=2.26.0